zigbee2mqtt/pyproject.toml
[project]
name = "zigbee2mqtt-tools"
version = "0.1.0"
description = "Local maintenance tools for Zigbee2MQTT (backup/restore, coordinator firmware helpers)."
requires-python = ">=3.9"
dependencies = [
  "cc2538-bsl",
  "paho-mqtt",
  "PyYAML",
]

[project.scripts]
z2m = "z2m_tools.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["z2m_tools"]
